nav를 header로 통합. 기본 Nav.js 제거
